feat: implement register user webpage
Most of the code has been taken from my React Jobly project:
https://github.com/myronchen-git/Springboard-SEC-Unit-41.2-React-Jobly-Frontend

Created the webpage / React component for registering a new user.  The
code was adapted from my React Jobly project:
https://github.com/myronchen-git/Springboard-SEC-Unit-41.2-React-Jobly-Frontend/blob/f8976ecdb56c9348e9982307de0fe7bcef7a9bad/src/routes/SignUp.jsx

main.jsx has been set up to utilize React Router's BrowserRouter, so
that React routes can be used.

A bare App component (in App.jsx) has been created with only the
registerUser function and display of the routes list.  It is bare,
because right now, the goal is to get a minimally functioning website.
The function is created here to help consolidate related code, for
processing, in one place.  The React component, in routes/Register.jsx,
doesn't directly call the API registerUser function, because that
should only be concerned with displaying the webpage.

The registerUser function has been created in ResumeManagerApi
(api.js).  This will serve as a convenient, self-explanatory, and
shorter function call.

The routes list has been created, with only the registration route for
now.  To keep it simple, necessary functions from App.jsx is currently
being passed to webpages through component props.  In the future, the
functions will be passed through a React context.  The purpose of the
routes list is to abstract out the routing paths, so that it is easier
to look at.

Tests have been skipped due to time, but manual tests have been
performed.