feat: implement update account webpage
Most of the code has been taken from my React Jobly project:
https://github.com/myronchen-git/Springboard-SEC-Unit-41.2-React-Jobly-Frontend

Created a React component called Account, which will be used as a
webpage, that will allow updating account-related info.  Currently, it
only allows updating the password.

Added the webpage to the routes list.

Added an updateAccount function in the App React component, so that the
logic is consolidated in one place.

An updateAccount function is created in ResumeManagerApi for
conveniently making an HTTP request to update an account.

Written tests have been skipped due to time, but manual tests have been
performed.  For manual testing, note that the app is a single page
application, so there is currently no way of navigating to different
pages without losing the stored authentication token.  This was worked
around by setting the catch-all route to navigate to the update account
webpage.