feat: implement NavBar React component
Created a navigation bar for going to the account webpage and different
webpages and signing out.  This only appears when signed in and is
togglable.

Updated RouteList React component to wrap NavBar for every route that
is used for signed-in users.  The wrapping, or nested routing, is done
to avoid rendering the NavBar for non-applicable routes, since the
NavBar will only contain links for signed-in users.

Created the signoutUser function in App React component, where the rest
of the user handling is.  The function is still being passed down to
the NavBar temporarily for now.  It will later be passed down using a
React context.

Written tests have been skipped for now due to time, but basic manual
tests have been performed.