feat: protect routes from unauthorized access
Created a component called RouteProtector and placed it as a route
element in the routes list, to act as middleware for routes that
require a logged-in user.  It will redirect or navigate the website app
to the home page if there is no user logged in.

For the Document React component, removed the temporary code in the
useEffect block that originally did the redirecting.

Written tests have been skipped due to time, but basic manual tests
have been performed.