feat: fill out functions in Document component
In the Document React component, the loadDocument function has been
fully implemented with code that retrieves the specified document and
its contents.

In the Document React component, the createDocument function has been
updated to also retrieve contact information, so that the new document
can display it without having to force the database to run a lengthy
SQL query to go through all the tables to retrieve contents that do not
exist yet.

For the createDocument function, the new document object is also added
to the list of already retrieved documents for the user, so that
another network call is not needed.  This will update the list of
documents for the user the next time the document selection form is
opened when staying on the same webpage.

These two changes for createDocument should improve performance.

Two methods are implemented in ResumeManagerApi to help with the new
functionality and are created as convenience methods.

Written tests have been skipped due to time, but basic manual tests
have been performed.