feat: implement add section in front-end
Created a React component, AddSectionCard, that renders a card/form for
adding a section to a document.  This will appear after all existing
sections.

The SectionsList React component is updated to include AddSectionCard.

SectionsList has a new function, addSection, which will do the handling
after a user submits the AddSectionCard form.  addSection is in
SectionsList, because it will make changes to the document object, and
those changes need to re-render the sections list.  The changes to the
document object is made locally and is done to save from having to make
a network call to get the document and all of its contents again.
Since the document object needs to be changed, SectionsList now has a
new parameter for setDocument.

SectionsList has a new effect that retrieves the available sections
that can be used in resumes.  The effect is here, because addSection
needs to use the retrieved sections to find the correct section names.

ResumeManagerApi now has two methods, getSections and addSection, to
help with the new functionality.  These methods are created for
convenience, which helps to reduce lengthy and repeated code.

Written tests have been skipped due to time, but basic manual tests
have been performed.