fix: display errors for document selection
For the DocumentSelect React component, it was not showing any errors
if the submission failed.  This can cause confusion for users and is
not good feedback.  For instance, if there is a network failure right
before a user selects a document, there would be no feedback.

Error messages are now being displayed on the component, below all
displayed documents.

Written tests have been skipped due to time, but basic manual tests
have been performed.