feat: implement EducationCard React component
Created the React component that will display an education.  Currently,
it shows the basic HTML for an education, but it still needs to later
allow editing and deleting, and it still needs styling.

Added EducationCard to the list of possible section item components to
render in SectionCard.

Written tests have been skipped due to time, but basic manual tests
have been performed.