fix: move onClick HTML attribute to right location
In React components, the onClick attributes/props were incorrectly
placed in the CardHeader components.  This caused the headers to be
clickable, when the icons were supposed to be the ones clickable.  This
has also resulted in HTTP errors, because the incorrect IDs have been
given, due to the "data-id" attributes not being found.  The onClick
attributes now have been moved to the img elements.

Consequently, finding the parent elements with "data-id" when clicking
on the icons had to be changed.  To make finding more reliable, instead
of traversing individual parents from an event target, the JavaScript
web API element method, "closest", is used.

Written tests have been skipped due to time, but basic manual tests
have been performed.