feat: implement delete experience functionality
Created front-end and back-end functionality for deleting an experience
and detaching it from a document.  The code is copied over from
education and will need to be refactored later.

In the back-end, created the routes and services for deleting an
experience entry and a document-experience relationship entry in the
database.

In the front-end, created a function, deleteExperience, in the
ExperienceCard React component to handle deleting an experience or
removing an experience from a document, depending on whether the
document is the master resume.

In the front-end, also created the related convenience methods.

Written tests have been skipped due to time, but basic manual tests
have been performed.  Since the code is from the education side, most
of it should be working.