feat: implement removing snippet from experience
Created front-end and back-end functionality for removing a text
snippet from an experience.  In other words, this deletes an
experience-text snippet relationship.

* Back-end:
In routes/experiences.js, implemented a new delete route for deleting
an experience-text snippet relationship, which needs document,
experience, and text snippet IDs.  Text snippet version is not needed,
because there can only be one text snippet of a particular ID that can
be associated with an experience in a document.  In other words, there
can not be two text snippets with the same ID but with different
versions.  This is because the experiences_x_text_snippets database
table only has the text snippet ID part as one of the primary keys.

In services/experienceService.js, implemented the corresponding service
function, called deleteExperience_x_textSnippet.  Only the document
ownership needed to be checked, because only experiences and text
snippets from the same owner can have an experience-text snippet
relationship.  If given another user's experience or text snippet ID,
the model functions called in deleteExperience_x_textSnippet would
throw a "not found" error or do nothing.

* Front-end:
In TextSnippetCard, called a generic function, detachTextSnippet, in
the deleteTextSnippet function to remove a text snippet, instead of
placing that code here.  The code to remove a text snippet from a
section has to be placed in experience, skill, or other components,
else all of the specific code for all of the sections would be in
TextSnippetCard.

In ExperienceCard, implemented the detachTextSnippet function, to
remove a text snippet from an experience in a particular document.
Note the code specific to removing a text snippet from experience.
This code would not be appropriate in TextSnippetCard.

In ResumeManagerApi, implemented the corresponding function,
removeTextSnippetFromExperience, to help make it easier to make the
corresponding API requests.

Written tests have been skipped due to time, but basic manual tests
have been performed.