refactor: utilize EducationForm in AddEducationCard
For the front-end, replaced redundant code in the AddEducationCard
React component by using EducationForm.

A new function, addEducation, is created to handle form data, pass the
data to the correct function, and update the document state.  This code
is specific to adding/creating an education and can not be in
EducationForm.

A new variable, initialFormData, is created to pass in the initial form
data to EducationForm.

The toggleOpen function is no longer needed, since there is no
interaction with the form data or error message states.  It is replaced
with direct calls to the setIsEducationFormOpen function.

Written tests have been skipped due to time, but basic manual tests
have been performed.