feat: implement form React component
Created the GenericForm React component, which will manage displaying
form inputs and will manage user inputs.  This is a more generalized
component than EducationForm and is meant to allow any section to use
it for adding or updating.

Note that since GenericForm is going to be used across sections and for
adding and editing, a unique name needs to be used for the input id.
This is currently being solved by using the timestamp, which will be
created every time the form is displayed, or opened.  There should be
no conflict in the ids between forms, because users would not be able
to instantaneous open more than two forms at once.

Written tests have been skipped due to time, but basic manual tests
have been performed.