feat: implement edit text snippet in front-end
In the TextSnippetCard React component, added the edit icon (which only
shows up in the master resume), HTML form for editing, and
editTextSnippet function.  The document state update could not be done
in editTextSnippet and had to be done in a parent section component,
because a text snippet could belong to an experience, skill, or project.

In ExperienceCard, implemented the replaceTextSnippetInDocumentState
function to update the document state after a text snippet is updated.
This allows for text snippets in experiences to be properly updated.

In commonData.js, added the text snippet fields constants.

In ResumeManagerApi, implemented the updateTextSnippet method to make
it easier to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.