feat: implement repositioning educations front-end
Implemented the functionality to allow dragging and dropping educations
within the education section and saving the new positions into the
back-end.  Implemented code have been adapted from examples from
@hello-pangea/dnd
(https://github.com/hello-pangea/dnd/blob/HEAD/docs/about/examples.md).

In SectionItemsList, added a @hello-pangea/dnd Droppable component to
the rendered output that wraps the list of section items, along with
passing the list index to section item components.  There is no new
DragDropContext, because nested DragDropContexts are not allowed.  The
list index is needed for Draggable components.  The Droppable component
is used to designate what HTML areas can have drag-and-drop.

In EducationCard, added a @hello-pangea/dnd Draggable component to the
rendered output to allow dragging educations.  The drag handle is a set
of dots, which is used to limit where on an education can initiate a
drag.

In SectionsList, implemented a function to reposition educations
visually and to send an API request to save the new positions to the
back-end.  This is similar to repositionSections and will need to be
refactored in a future commit.  Also, in the onDragEnd function, added
conditions to select the correct repositioning function according to
type, though this may be removed in the future refactor.

All Droppables are given types, so that dragging and dropping
educations does not interact with sections, and vice versa.

In ResumeManagerApi, implemented the repositionEducations method to make
it easier to call the back-end API.

In resume-manager-backend\routes\educations.js, additionally fixed up
the route description in the back-end for the PUT HTTP request method
for updating education positions.

Written tests have been skipped due to time, but basic manual tests
have been performed.