feat: implement reposition experiences front-end
Implemented the functionality to allow dragging and dropping experiences
within the experience section and saving the new positions into the
back-end.

In ExperienceCard, implemented the Draggable component from
@hello-pangea/dnd and added the dots icon as the drag handle.

In SectionsList, added handling for experiences in the onDragEnd
function.

In ResumeManagerApi, implemented the repositionExperiences method to
make it easier to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.