perf: limit GET calls when attaching in front-end
In AttachEducationCard, AttachExperienceCard, and
AttachTextSnippetCard, there is an API request / network call that
occurs every time when those components are mounted.  This network call
may be unnecessary as users might not want to attach a particular
education, experience, or text snippet to the current document.

To reduce unnecessary calls, the call has been moved from the useEffect
block and into the toggleOpen function.  The call now only occurs when
users first open the attachment form.

The availableEducations and availableExperiences states have also been
initialized to null instead of an empty Array to reflect when a network
call has been made to retrieve the available educations and experiences
and when a network call has been successfully made but there are no
available items.

Written tests have been skipped due to time, but basic manual tests
have been performed.