refactor: consolidate document form code
In the front-end, the NewDocumentForm and EditDocumentForm React
components shared almost exactly the same code, therefore it has been
extracted into a new general component, DocumentForm.  So far,
DocumentForm should only be used for creating a new document or edit an
existing one.

One major logic has been changed: the rendering of the isTemplate and
isLocked fields now depend on if the initial form data passed into
DocumentForm has those properties.  This was done to simplify the logic.

In the Document component, NewDocumentForm and EditDocumentForm have
been replaced with DocumentForm.  The initial form data is created
within the component output (return statement), which should be fine,
because the Document should not be going through many re-renders while
users interact with the DocumentForm.  The alternative would be to use
useMemo and useCallback for all of the props of DocumentForm.

Written tests have been skipped due to time, but basic manual tests
have been performed.