fix: properly update document state for text snips
In the ExperienceCard React component, the addTextSnippet and
attachTextSnippet functions were improperly updating the document
state.  React suggests not to mutate props, contexts, and states.

The code that updates the document state with a new text snippet for an
experience has been rewritten and abstracted out to a function called
addTextSnippetToDocumentState.  In that function, the document,
experience, and experience bullets are properly copied and the
originals have not been mutated.

Written tests have been skipped due to time, but basic manual tests
have been performed.