fix: remove wide Navbar Reactstrap component
In the front-end's NavBar component's rendered output, the Navbar
Reactstrap component was used with the "fixed" option, which created a
an element that spanned across the viewport width.  This blocked the
buttons at the top of the document webpage, making the top 90% of the
buttons unclickable.

The Navbar has been removed and CSS has been applied to the button
instead.

Also, adding CSS in the NavBar JSX code was not necessary after all.
The CSS selector in the related CSS file just had to be more specific,
so that it has precedence over Bootstrap's CSS.