feat(front-end): stylize sections
Added margins to card Reactstrap components.  It had to be applied to
the draggable attribute, otherwise the other cards will move when
dragging a card around.

Made the card headers smaller to reduce space.

In the SectionsList React component, removed the extra article element
in the rendered output.  It wrapped around the Droppable component and
all sections, which did not make sense when each SectionCard component
was supposed to be an article element.