feat(front-end): stylize education text
Created a React component, called EducationText, to abstract out the
rendering of an education's text, because it was not simple.  The
user-friendly date Strings were created using React's useMemo, because
converting a String to a Date and back to a formatted String may be an
expensive process.  The portion of the education text related to awards
and activities utilized useMemo, because there are a good amount of
statements used to create that text.  useMemos also were utilized
because the education text should not be changing too much.

In EducationCard, replaced the old displaying of education info with a
component called EducationText.  The new display of the info required
using a CSS grid, because it was the easiest way of having two elements
side-by-side while having a third one underneath with the same width of
the two combined.  To easily apply the CSS grid to only when displaying
education info and not when displaying the edit form, conditional
classes were set for the CardBody component.