refactor: generalize and move styling for education
The styling applied to education cards could also be applied to
experience cards, so it has been moved to the nearest parent React
component of the two, which is SectionItemsList.  Any HTML class names
specific to education that are used to apply the CSS have been
generalized.

Also made the third part of the card have a width of 100%, so that text
snippet cards are not small.