feat(front-end): stylize experience text
Created a React component, called ExperienceText, to abstract out the
rendering of an experience's text and reduce clutter in
ExperienceCard.  The user-friendly date Strings were created using
React's useMemo, because converting a String to a Date and back to a
formatted String may be an expensive process.

In ExperienceCard, conditionally applied the class for the card CSS
styling to the CardBody Reactstrap component, so that it does not get
applied to the edit form and because CardBody is the containing HTML
element for the experience text.  Also replaced the placeholder JSX for
displaying experience info with ExperienceText.

The HTML class for styling the third part of a card had to be applied
to TextSnippetsList, because there was no way of setting the
"section-item-text__p3" class name in ExperienceText.

Moved the generateDateString function in EducationText out into a
utility file called conversions.js, because ExperienceText also needs
to use it.

In Document.css, set the default text alignment to center for cards in
the element with HTML ID "Document", because the "Add Text +" button
was not centered and so that other elements are centered by default as
well.