fix: align form buttons
The submit button on some forms, such as the one for editing a text
snippet was aligned to the left.  The GenericForm React component has
been given a CSS property to align its children to the center, which
should have been the default.

The submit button on some forms was also too close to the "select"
element above.  All "select" elements in "form" elements and card
components are now given bottom margins.