chore: commit various general document styling
Increased the z-index CSS for the menu button on the document webpage,
so that it does not get covered by the document.

Added a margin below section card titles.

Added padding to the document webpage.

Displayed document name above the document on the document webpage, so
users know which document they are working on.